+++ girepository/gitypelib.c
@@ -87,7 +87,7 @@ get_dir_entry_checked (GITypelib *typelib,
       return FALSE;
     }
 
-  offset = header->directory + (index - 1) * header->entry_blob_size;
+  offset = header->directory + (index - 1u) * header->entry_blob_size;
 
   if (typelib->len < offset + sizeof (DirEntry))
     {
@@ -161,7 +161,8 @@ gi_typelib_get_dir_entry (GITypelib *typelib,
 {
   Header *header = (Header *)typelib->data;
 
-  return (DirEntry *)&typelib->data[header->directory + (index - 1) * header->entry_blob_size];
+  /* this deliberately doesn’t check for underflow of @index; see get_dir_entry_checked() for that */
+  return (DirEntry *)&typelib->data[header->directory + (index - 1u) * header->entry_blob_size];
 }
 
 static Section *
@@ -303,7 +304,7 @@ strsplit_iter_next (StrSplitIter  *iter,
   if (next)
     {
       iter->s = next + iter->sep_len;
-      len = next - s;
+      len = (size_t) (next - s);
     }
   else
     {
@@ -1718,7 +1719,7 @@ validate_object_blob (ValidateContext *ctx,
     }
 
   if (typelib->len < offset + sizeof (ObjectBlob) +
-            (blob->n_interfaces + blob->n_interfaces % 2) * 2 +
+            (blob->n_interfaces + blob->n_interfaces % 2u) * 2u +
             blob->n_fields * sizeof (FieldBlob) +
             blob->n_properties * sizeof (PropertyBlob) +
             blob->n_methods * sizeof (FunctionBlob) +
@@ -1875,7 +1876,7 @@ validate_interface_blob (ValidateContext *ctx,
     return FALSE;
 
   if (typelib->len < offset + sizeof (InterfaceBlob) +
-            (blob->n_prerequisites + blob->n_prerequisites % 2) * 2 +
+            (blob->n_prerequisites + blob->n_prerequisites % 2u) * 2u +
             blob->n_properties * sizeof (PropertyBlob) +
             blob->n_methods * sizeof (FunctionBlob) +
             blob->n_signals * sizeof (SignalBlob) +
